requirements.txt
pyside6
python-docx
pdfplumber